准备切到另外一个 广告图空间

+++ project1/widget/AOImageView.h
@@ -10,7 +10,7 @@
 //按钮点击协议
 @protocol UrLImageButtonDelegate <NSObject>
 
--(void)click:(int)vid;
+-(void)click:(NSInteger)vid;
 
 @end
